code: rename DetoxInvariantError to DetoxInternalError
